added parkingFee calculation for reservations

+++ src/main/java/ch/uzh/ifi/hase/soprafs22/entity/Parkingslip.java
@@ -46,7 +46,7 @@ public class Parkingslip implements Serializable {
     private String licensePlate;
 
     @Column(nullable = true)
-    private long parkingFee;
+    private float parkingFee;
 
     public long getId() {
         return id;
@@ -112,11 +112,11 @@ public class Parkingslip implements Serializable {
         this.licensePlate = licensePlate;
     }
 
-    public long getParkingFee() {
+    public float getParkingFee() {
         return parkingFee;
     }
 
-    public void setParkingFee(long parkingFee) {
+    public void setParkingFee(float parkingFee) {
         this.parkingFee = parkingFee;
     }
