+++ src/main/java/ch/uzh/ifi/hase/soprafs22/rest/dto/ReservationGetDTO.java
@@ -10,7 +10,7 @@ public class ReservationGetDTO {
     private String checkoutDate;
     private String checkoutTime;
     private String licensePlate;
-    private long parkingFee;
+    private float parkingFee;
 
     public long getReservationId() {
         return reservationId;
@@ -76,11 +76,11 @@ public class ReservationGetDTO {
         this.licensePlate = licensePlate;
     }
 
-    public long getParkingFee() {
+    public float getParkingFee() {
         return parkingFee;
     }
 
-    public void setParkingFee(long parkingFee) {
+    public void setParkingFee(float parkingFee) {
         this.parkingFee = parkingFee;
     }
 }
